+++ lib/command/impl/run/run.dart
@@ -14,9 +14,9 @@ class Run extends Command {
 
   @override
   ArgParser get argParser {
-    var argParser = ArgParser();
+    final argParser = ArgParser();
     argParser.addSeparator('Usage: yuro run <package>');
-    var sb = StringBuffer()..writeln('Available packages:');
+    final sb = StringBuffer()..writeln('Available packages:');
     _commands.forEach((element) {
       sb.write(element.name);
       sb.write(' ' * (15 - element.name.length));
@@ -31,10 +31,10 @@ class Run extends Command {
   Future<void> parser(List<String> arguments) async {
     if (arguments.isNotEmpty) {
       if (argParser.commands.containsKey(arguments[0])) {
-        var command = _commands.where((element) => element.name == arguments[0]).first;
+        final command = _commands.where((element) => element.name == arguments[0]).first;
         await command.parser(arguments.sublist(1));
       } else {
-        throw ArgumentError('Could not find a command named "${arguments[0]}".');
+        logger.e('Could not find a command named "${arguments[0]}".');
       }
     } else {
       stdout.writeln(argParser.usage);